META-RAT - update -big-difrent/requirements.txt
socket
plyer
wallpaper
subprocess
platform
psutil
pyautogui
pycaw.pycaw
comtypes
ctypes
pyttsx3
os
getpass import getuser
winsound
pynput 
pynput.keyboard
logging
PyWallpaper
struct
sqlite3
ctypes
requests
time
shutil
pyAesCrypt
http.server
pymsgbox
colorama